+++ core/main.ts
@@ -14,11 +14,12 @@ export * from "./src/contrib/systems/core/NPCSystem";
 export * from "./src/contrib/systems/core/PhysicsSystem";
 export * from "./src/contrib/systems/gameplay/EnemySpawnSystem";
 export * from "./src/contrib/systems/gameplay/PlayerSpawnSystem";
-export * from "./src/contrib/systems/net/TickData";
 export * from "./src/contrib/systems/net/delay/DelayClientSystem";
 export * from "./src/contrib/systems/net/delay/DelayServerSystem";
+export * from "./src/contrib/systems/net/delay/DelayTickData";
 export * from "./src/contrib/systems/net/rollback/RollbackClientSystem";
 export * from "./src/contrib/systems/net/rollback/RollbackServerSystem";
+export * from "./src/contrib/systems/net/rollback/RollbackTickData";
 export * from "./src/contrib/systems/ui/ChatBubbleSystem";
 export * from "./src/contrib/systems/ui/ClickableSystem";
 export * from "./src/contrib/systems/ui/DebugSystem";
