+++ core/src/contrib/components/actions/PlayerControlsEntity.ts
@@ -3,7 +3,7 @@ import { Action, Controlled, Controlling, Entity } from "@piggo-gg/core";
 export const PlayerControlsEntity: Action = Action(({ entity, world, player }) => {
   if (!entity || !player) return;
 
-  // check that the entity isn't already being controlled
+  // check that the entity isn't already controlled
   if (entity.components.controlled) return;
 
   // get the player entity
