+++ core/src/contrib/systems/net/rollback/RollbackClientSystem.ts
@@ -1,4 +1,4 @@
-import { InvokedAction, Entity, SystemBuilder, TickData, World } from "@piggo-gg/core";
+import { InvokedAction, Entity, SystemBuilder, RollbackTickData, World, Noob, Ball, Skelly, Zombie } from "@piggo-gg/core";
 
 const servers = {
   dev: "ws://localhost:3000",
@@ -6,7 +6,7 @@ const servers = {
   production: "wss://api.piggo.gg"
 } as const;
 
-// WssNetcodeSystem handles networked entities over WebSockets
+// rollback netcode client
 export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
   id: "RollbackClientSystem",
   init: ({ world, clientPlayerId }) => {
@@ -21,10 +21,10 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
     }, 200);
 
     let lastMessageTick: number = 0;
-    let latestServerMessage: TickData | null = null;
+    let latestServerMessage: RollbackTickData | null = null;
 
     wsClient.onmessage = (event) => {
-      const message = JSON.parse(event.data) as TickData;
+      const message = JSON.parse(event.data) as RollbackTickData;
 
       // ignore messages from the past
       if (latestServerMessage && (message.tick < latestServerMessage.tick)) return;
@@ -80,7 +80,7 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
         }
       });
 
-      const message: TickData = {
+      const message: RollbackTickData = {
         type: "game",
         tick: world.tick,
         timestamp: Date.now(),
@@ -96,11 +96,11 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
     const handleLatestMessage = () => {
       if (latestServerMessage === null) return;
       let message = latestServerMessage;
-      let rollback = false;
+      let shouldRollback = false;
 
       const mustRollback = (log: string) => {
-        if (!rollback) {
-          rollback = true;
+        if (!shouldRollback) {
+          shouldRollback = true;
           console.log(`rollback from ${message.tick}! ${log}`);
         }
       }
@@ -108,7 +108,7 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
       if (message.tick > world.tick) mustRollback("server is ahead");
 
       // handle future actions
-      if (!rollback) {
+      if (!shouldRollback) {
         Object.keys(message.actions).map(Number).filter((t) => t > world.tick).forEach((futureTick) => {
           Object.keys(message.actions[futureTick]).forEach((entityId) => {
             if ((entityId === clientPlayerId) || (entityId === `skelly-${clientPlayerId}`)) return;
@@ -118,7 +118,7 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
       }
 
       // compare action buffers
-      if (!rollback) Object.keys(message.actions).map(Number).forEach((tick) => {
+      if (!shouldRollback) Object.keys(message.actions).map(Number).forEach((tick) => {
 
         if (tick < message.tick) return;
 
@@ -151,14 +151,14 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
       });
 
       // compare entity counts
-      if (!rollback && world.entitiesAtTick[message.tick]) {
+      if (!shouldRollback && world.entitiesAtTick[message.tick]) {
         if (Object.keys(world.entitiesAtTick[message.tick]).length !== Object.keys(message.serializedEntities).length) {
           mustRollback(`entity count local:${Object.keys(message.serializedEntities).length} remote:${Object.keys(world.entitiesAtTick[message.tick]).length}`);
         }
       }
 
       // compare entity states
-      if (!rollback) {
+      if (!shouldRollback) {
         Object.entries(message.serializedEntities).forEach(([entityId, msgEntity]) => {
           const entitiesAtTick = world.entitiesAtTick[message.tick];
           if (entitiesAtTick) {
@@ -187,7 +187,7 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
         });
       }
 
-      if (rollback) world.rollback(message);
+      if (shouldRollback) rollback(world, message);
     }
 
     return {
@@ -198,3 +198,71 @@ export const RollbackClientSystem: SystemBuilder<"RollbackClientSystem"> = ({
     }
   }
 });
+
+const rollback = (world: World, td: RollbackTickData) => {
+  const now = Date.now();
+
+  // determine how many ticks to increment
+  world.framesAhead = Math.ceil((((world.ms) / world.tickrate) * 2) + 1);
+  if (Math.abs(world.framesAhead - (world.tick - td.tick)) <= 1) {
+    world.framesAhead = world.tick - td.tick;
+  }
+
+  console.log(`ms:${world.ms} msgFrame:${td.tick} clientFrame:${world.tick} targetFrame:${td.tick + world.framesAhead}`);
+
+  // set tick
+  world.tick = td.tick - 1;
+
+  // remove old local entities
+  Object.keys(world.entities).forEach((entityId) => {
+    if (world.entities[entityId].components.networked) {
+
+      if (!td.serializedEntities[entityId]) {
+        // delete if not present in rollback frame
+        console.log("DELETE ENTITY", entityId, td.serializedEntities);
+        world.removeEntity(entityId);
+      }
+    }
+  });
+
+  // add new entities if not present locally
+  Object.keys(td.serializedEntities).forEach((entityId) => {
+    if (!world.entities[entityId]) {
+      if (entityId.startsWith("zombie")) {
+        world.addEntity(Zombie({ id: entityId }));
+      } else if (entityId.startsWith("ball")) {
+        world.addEntity(Ball({ id: entityId }));
+      } else if (entityId.startsWith("noob")) {
+        world.addEntity(Noob({ id: entityId }))
+      } else if (entityId.startsWith("skelly")) {
+        world.addEntity(Skelly(entityId));
+      } else {
+        console.error("UNKNOWN ENTITY ON SERVER", entityId);
+      }
+    }
+  });
+
+  // deserialize everything
+  Object.keys(td.serializedEntities).forEach((entityId) => {
+    if (world.entities[entityId]) {
+      world.entities[entityId].deserialize(td.serializedEntities[entityId]);
+    }
+  });
+
+  // update local action buffer
+  Object.keys(td.actions).map(Number).forEach((tick) => {
+    Object.keys(td.actions[tick]).forEach((entityId) => {
+      // skip future actions for controlled entities
+      if (tick > td.tick && world.entities[entityId]?.components.controlled?.data.entityId === world.clientPlayerId) return;
+
+      world.actionBuffer.set(tick, entityId, td.actions[tick][entityId]);
+    });
+  });
+
+  Object.values(world.systems).forEach((system) => system.onRollback ? system.onRollback() : null);
+
+  // run system updates
+  for (let i = 0; i < world.framesAhead + 1; i++) world.onTick({ isRollback: true });
+
+  console.log(`rollback took ${Date.now() - now}ms`);
+}
