+++ core/src/contrib/systems/net/rollback/RollbackTickData.ts
@@ -1,6 +1,6 @@
 import { InvokedAction, SerializedEntity } from "@piggo-gg/core";
 
-export type TickData = {
+export type RollbackTickData = {
   type: "game"
   tick: number
   timestamp: number
@@ -8,16 +8,5 @@ export type TickData = {
   player: string
   actions: Record<number, Record<string, InvokedAction[]>>
   chats: Record<number, Record<string, string[]>>
-  serializedEntities: Record<string, SerializedEntity>
-}
-
-export type DelayTickData = {
-  type: "game"
-  tick: number
-  timestamp: number
-  latency?: number
-  player: string
-  actions: Record<string, InvokedAction[]>
-  chats: Record<string, string[]>
   serializedEntities: Record<string, SerializedEntity>
 }
